Create infograb.bat
This script is used to run HijackThis and Tcpview (via command line util tcpvcon.exe). You only need to have tcpvcon.exe and hijackthis.exe (doesn't have to be case sensitive, HAS to be named in this fashion). At the end of the script you will have two log files.